Bluetooth first testing

+++ PlatformIO_codebase/include/PIDController.h
@@ -5,12 +5,12 @@ class PIDController {
 private:
     float Kp, Ki, Kd;
     float integral;
-    float previousError;
+    double previousError;
     unsigned long lastTime;
 
 public:
     PIDController(float p, float i, float d);
-    float compute(float error);
+    double compute(float error);
     void reset();  // clear integral/prev error
     void setGains(float p, float i, float d);  // update PID gains dynamically
 };
